+++ db/migrate/20200814210142_add_tip_to_orders.rb
@@ -1,5 +1,5 @@
 class AddTipToOrders < ActiveRecord::Migration[6.0]
   def change
-    add_column :orders, :tip, :float
+    add_column :orders, :tip, :integer
   end
 end
